Move PdfDocument to root
